services/cockpit-core/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
celery[redis]==5.4.0
httpx==0.27.2
pydantic-settings==2.6.1
redis==5.1.1
psycopg[binary]==3.2.1
